// FrontEnd/src/app/models/fornecedor.ts
export interface Fornecedor {
  id: number;
  nome: string;
  fantasia: string;
  endereco: string;
  numero: string;
  complemento: string;
  bairro: string;
  uf: number;
  cidade: number;
  cep: string;
  telefone_1: string;
  telefone_2: string;
  email: string;
  cnpj: string;
  cpf: string;
  rg: string;
  nomeuf: string;
  nomecidade: string;
}